Move all csv logic into custom hook
